// grayjay-android/app/src/unstable/assets/sources/bilibili/src/types.ts
export type RequestMetadata<X> = {
    request(builder: BatchBuilder): BatchBuilder
    process(http_response: BridgeHttpResponse<string>): X
}
// ...
export type Settings = {
    readonly verboseNotifications?: boolean | string
    readonly multipartNextOnlyRecommendations?: boolean | string
}

export type PluginSettings = {
    readonly verboseNotifications?: boolean
    readonly multipartNextOnlyRecommendations?: boolean
}
// ...
export type Wbi = { readonly wbi_img_key: string, readonly wbi_sub_key: string }

export type State = {
    readonly buvid3: string
    readonly buvid4: string
    readonly buvid_fp: string
    readonly b_nut: number
    readonly mixin_key: string
    readonly dm_cover_img_str: string
    readonly dm_img_str: string
    readonly dm_img_inter: string
}
// ...
export type FingerSpiResponse = {
    readonly data: {
        readonly b_3: string
        readonly b_4: string
    }
}

export type NavResponse = {
    readonly data: {
        readonly isLogin: false
        readonly wbi_img: {
            readonly img_url: string,
            readonly sub_url: string
        }
    }
}